+++ daisi/src/cpps/agv/agv_logical.h
@@ -25,9 +25,9 @@
 #include "cpps/amr/message/amr_order_info.h"
 #include "cpps/amr/message/amr_order_update.h"
 #include "cpps/amr/message/amr_status_update.h"
+#include "cpps/amr/model/amr_static_ability.h"
 #include "cpps/common/agv_description.h"
 #include "cpps/common/cpps_logger_ns3.h"
-#include "cpps/model/ability.h"
 #include "cpps/model/kinematics.h"
 #include "cpps/model/task.h"
 #include "cpps/negotiation/mrta_configuration.h"
