+++ daisi/src/cpps/amr/model/amr_load_carrier.h
