+++ daisi/src/cpps/amr/model/amr_static_ability.cpp
